Fix: yt-dlp - HTTP Error 403: Forbidden. Closes #161

+++ pkg/utils/bytes.go
@@ -9,7 +9,7 @@ const (
 	TB = GB * 1024
 )
 
-func BytesToHuman(b int) string {
+func BytesToHuman(b int64) string {
 	switch {
 	case b >= TB:
 		return fmt.Sprintf("%.2f TiB", float64(b)/float64(TB))
